implement hit button functionality
This is the first in a series of commits that will add the controller and routes
as a part of this application. There are some minor front and backend changes.
I'm choosing to commit here because having 'hit' working seemed like a good
milstone. The next commits will continue to add functionality to the buttons.